% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition_plot.R
\name{composition_plot}
\alias{composition_plot}
\title{This function returns a plot composition representing the absolute number 
and proportion of a variable grouped by a feature of interest.}
\usage{
composition_plot(
  metadata,
  x,
  grouping_vr,
  gradient_col = c("#fffedb", "#ffc157", "#fa5e25"),
  col_df = NULL,
  x_order = NULL,
  grouping_order = NULL,
  title = "",
  label = FALSE
)
}
\arguments{
\item{metadata}{metadata from the Seurat object - se_obj\@meta.data}

\item{x}{character string indicating the X axis - cell type annotation}

\item{grouping_vr}{character string specifying the feature to group
the data by - condition}

\item{gradient_col}{vector of n colors indicating the color gradient to
use in the proportion tile plot}

\item{col_df}{data frame with variables name and color containing
the name - color pair}

\item{x_order}{character string vector containing the x names in the
desired order}

\item{grouping_order}{character string vector containing the grouping names
in the desired order}

\item{title}{character string indicating the plot title, by default and
empty string ""}

\item{label}{Logical vector indicating if we want to show proportion labels
in the tile plot. By default FALSE.}
}
\value{
Composition plot
}
\description{
This function returns a plot composition representing the absolute number 
and proportion of a variable grouped by a feature of interest.
}
\examples{
\dontrun{
library(Seurat)
library(SeuratData)
# SeuratData::InstallData(ds = "pbmcsca")
se_obj <- SeuratData::LoadData(ds = "pbmcsca")

col_df <- data.frame(
  "name" = unique(se_obj@meta.data$CellType),
  "color" = terrain.colors(10)
)

grouping_order <- c("Cytotoxic T cell", "CD4+ T cell", "CD14+ monocyte", "B cell",
             "Megakaryocyte", "CD16+ monocyte", "Natural killer cell",
             "Dendritic cell", "Plasmacytoid dendritic cell", "Unassigned")
SCrafty::composition_plot(
  metadata = se_obj@meta.data,
  x = "Method",
  grouping_vr = "CellType",
  gradient_col = rainbow(5),
  col_df = col_df,
  x_order = unique(se_obj@meta.data$Method),
  grouping_order = grouping_order,
  title = "Distribution of cells over methods"
)

}

}
