% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_geneid.R
\name{convert_geneid}
\alias{convert_geneid}
\title{This function converts a vector of gene identifiers to other identifiers}
\usage{
convert_geneid(
  gene_vec,
  gene_from = "SYMBOL",
  gene_to = "ENTREZID",
  annotation = "org.Hs.eg.db"
)
}
\arguments{
\item{gene_vec}{Vector of gene identifiers to convert to
entrezID.}

\item{gene_from}{Character string specifying the gene identifier is gene_vec,
By default SYMBOL, can also be ENTREZID, ENSEMBL, UNIGENE}

\item{gene_to}{Character string specifying what the gene id we want to
convert the gene_vec to. By default ENTREZID, can also be SYMBOL, ENSEMBL,
UNIGENE}

\item{annotation}{Character string specifying specifying the annotation of
the organism to use. By default human org.Hs.eg.db but you can also choose
mouse org.Mm.eg.db.}
}
\value{
Named vector with ENTREZID as the values.
}
\description{
This function converts a vector of gene identifiers to other identifiers
}
\examples{
\dontrun{
  symbol_vec <- c("MS4A1", "CD4", "FOXP3")
  entrezid_vec <- convert_geneid(
  gene_vec = symbol_vec,
  gene_from = "SYMBOL",
  gene_to = "ENTREZID",
  annotation = "org.Hs.eg.db")
}
}
