% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_symb_entrez.R
\name{convert_symb_entrez}
\alias{convert_symb_entrez}
\title{This function converts a vector of gene symbols to its entrezID form so that
it can be used to perform Go enrichment analysis.}
\usage{
convert_symb_entrez(
  gene_vec,
  gene_type = c("SYMBOL", "ENSEMBL"),
  annotation = c("org.Hs.eg.db", "org.Mm.eg.db")
)
}
\arguments{
\item{gene_vec}{Vector of gene symbols or ensemble identifier to convert to entrezID.}

\item{gene_type}{Character string specifying if the gene vector are SYMBOL or ENSEMBLE identifiers. By default SYMBOL, can also be ENSEMBL.}

\item{annotation}{Character string specifying specifying the annotation of the organism to use. By default human org.Hs.eg.db but you can also choose mouse org.Mm.eg.db.}
}
\value{
Named vector with ENTREZID as the values.
}
\description{
Saves the plot in a tmp location and loads it as an image.
}
\examples{
\dontrun{
  symbol_vec <- c("MS4A1", "CD4", "FOXP3")
  entrezid_vec <- convert_symb_entrez(
  gene_vec = symbol_vec,
  gene_type = "SYMBOL",
  annotation = "org.Hs.eg.db")
}
}
