% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_matrix.R
\name{correlation_heatmap}
\alias{correlation_heatmap}
\title{This function returns a correlation heatmap matrix between the features of
interest specified including genes and or metadata features.}
\usage{
correlation_heatmap(
  se,
  genes = NULL,
  feats = NULL,
  assay = "Spatial",
  slot = "data",
  cor_method = "pearson",
  ...
)
}
\arguments{
\item{se}{Spatial Seurat object we want to assess.}

\item{genes}{vector with genes that we want to build
the correlation heatmap with.}

\item{feats}{vector with metadata features that we want to build
the correlation heatmap with.}

\item{assay}{assay from where we want to extract the gene expression.}

\item{slot}{slot from where we want to extract the gene expression.}

\item{cor_method}{Method to use for correlation: 
c("pearson", "kendall", "spearman"). By default pearson.}

\item{...}{Further arguments to pass to ggcorrplot::ggcorrplot}
}
\value{
Correlation matrix heatmap
}
\description{
This function returns a correlation heatmap matrix between the features of
interest specified including genes and or metadata features.
}
\examples{
\dontrun{
library(Seurat)
library(SeuratData)
sp_obj <- SeuratData::LoadData(ds = "stxBrain", type = "posterior1")
correlation_heatmap(
  se = sp_obj,
  slot = "counts",
  assay = "Spatial",
  genes = c("Ms4a1", "Cd79a", "Cd3d", "Cd3e", "Cd8a"),
  feats = NULL,
  cor_method = "pearson",
  title = "Gene-Gene correlation heatmap")
}

}
