% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpreview.R
\name{ggpreview}
\alias{ggpreview}
\title{This function allows you to see plots inline, especially useful when trying
to format font sizes.}
\usage{
ggpreview(x, w = 5, h = 5, dpi = 150, units = "in")
}
\arguments{
\item{x}{plot to visualize}

\item{w}{width in inches}

\item{h}{height in inches}

\item{dpi}{resolution}

\item{units}{character string specifying the units of the dimensions.
By default in, inches, but it can also be px, cm, or mm.}
}
\value{
Show plot image in Viewer pane
}
\description{
Saves the plot in a tmp location and loads it as an image.
}
\examples{
\dontrun{
library(ggplot2)
df <- data.frame(x = 1:10, y = 1:10)
tmp_plt <- ggplot2::ggplot(df, ggplot2::aes(x = x, y = y)) +
ggplot2::geom_point()
ggpreview(x = tmp_plt, w = 9, h = 4, units = "in")
}
}
