% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_gene_orgs.R
\name{human_gene_to_mouse}
\alias{human_gene_to_mouse}
\title{Basic function to convert human to mouse gene names}
\usage{
human_gene_to_mouse(x)
}
\arguments{
\item{x}{Human gene symbol list to convert to mouse}
}
\value{
Converted gene list
}
\description{
Basic function to convert human to mouse gene names
}
\examples{
\dontrun{
library(Seurat)
# A list of cell cycle markers, from Tirosh et al, 2015, is loaded with Seurat.  We can
# segregate this list into markers of G2/M phase and markers of S phase
s.genes_hs <- cc.genes$s.genes

s.genes_mm <- human_gene_to_mouse(x = s.genes_hs)

}

}
