% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_variance_plot.R
\name{mean_variance_plot}
\alias{mean_variance_plot}
\title{This function returns 2 plots to assess the mean-variance relationship of the
count data.}
\usage{
mean_variance_plot(se_obj, assay, slot)
}
\arguments{
\item{se_obj}{Seurat object we want to assess.}

\item{assay}{assay from where we want to extract the count matrix.}

\item{slot}{slot from where we want to extract the count matrix.}
}
\value{
patchwork object with 2 plots
}
\description{
This function returns 2 plots to assess the mean-variance relationship of the
count data.
}
\examples{
\dontrun{
library(Seurat)
library(dplyr)
data("pbmc_small")
mean_variance_plot(se_obj = pbmc_small, assay = "RNA", slot = "counts")
}
}
