% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_spatial.R
\name{mit_pct_hist_st}
\alias{mit_pct_hist_st}
\title{This function returns the mitochondrial \% histogram of the Seurat object.}
\usage{
mit_pct_hist_st(se, mito_percent = "percent.mito", sample_id = "orig.ident")
}
\arguments{
\item{se}{Spatial Seurat object we want to assess.}

\item{mito_percent}{Character string indicating the name of the variable in
the metadata containing the mitochondrial \% for each spot.}

\item{sample_id}{Feature in metadata by which to identify the sample if the
object is merged and contains several samples.}
}
\value{
Histogram with the mitochondrial \% distribution by sample.
}
\description{
This function returns the mitochondrial \% histogram of the Seurat object.
}
\examples{
\dontrun{
library(Seurat)
library(SeuratData)
sp_obj <- SeuratData::LoadData(ds = "stxBrain", type = "posterior1")
mit_pct_hist_st(se = sp_obj,
                 mito_percent = "percent.mito",
                 sample_id = "orig.ident")
}

}
