% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_gene_orgs.R
\name{mouse_gene_to_human}
\alias{mouse_gene_to_human}
\title{Basic function to convert mouse to human gene names}
\usage{
mouse_gene_to_human(x)
}
\arguments{
\item{x}{Mouse gene symbol list to convert to human}
}
\value{
Converted gene list
}
\description{
Basic function to convert mouse to human gene names
}
\examples{
\dontrun{
library(Seurat)
# A list of cell cycle markers, from Tirosh et al, 2015, is loaded with Seurat.  We can
# segregate this list into markers of G2/M phase and markers of S phase
musGenes <- c("Hmmr", "Tlx3", "Cpeb4")

hsGenes <- convertMouseGeneList(x = musGenes)

}

}
