% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_spatial.R
\name{ncount_hist_st}
\alias{ncount_hist_st}
\title{This function returns the number of counts histogram of the Seurat object.}
\usage{
ncount_hist_st(se, ncount = "nCount_Spatial", sample_id = "orig.ident")
}
\arguments{
\item{se}{Spatial Seurat object we want to assess.}

\item{sample_id}{Feature in metadata by which to identify the sample if the
object is merged and contains several samples.}

\item{ncounts}{Character string indicating the name of the variable in the
metadata containing the number of genes.}
}
\value{
Histogram with the nUMI distribution by sample.
}
\description{
This function returns the number of counts histogram of the Seurat object.
}
\examples{
\dontrun{
library(Seurat)
library(SeuratData)
sp_obj <- SeuratData::LoadData(ds = "stxBrain", type = "posterior1")
ncount_hist_st(se = sp_obj,
                 ncount = "nCount_Spatial",
                 sample_id = "orig.ident")
}

}
