% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_spatial.R
\name{nfeature_hist_st}
\alias{nfeature_hist_st}
\title{This function returns the number of features histogram of the Seurat object.}
\usage{
nfeature_hist_st(se, nfeat = "nFeature_Spatial", sample_id = "orig.ident")
}
\arguments{
\item{se}{Spatial Seurat object we want to assess.}

\item{nfeat}{Character string indicating the name of the variable in the
metadata containing the number of genes.}

\item{sample_id}{Feature in metadata by which to identify the sample if the
object is merged and contains several samples.}
}
\value{
Histogram with the nFeature distribution by sample.
}
\description{
This function returns the number of features histogram of the Seurat object.
}
\examples{
\dontrun{
library(Seurat)
library(SeuratData)
sp_obj <- SeuratData::LoadData(ds = "stxBrain", type = "posterior1")
nfeature_hist_st(se = sp_obj,
                 nfeat = "nFeature_Spatial",
                 sample_id = "orig.ident")
}

}
