% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_viz.R
\name{palette_viz}
\alias{palette_viz}
\title{This function allows you to visualize a palette}
\usage{
palette_viz(col_vec)
}
\arguments{
\item{col_vec}{vector containing the palette colors}
}
\value{
Tile plot showing the palette composition
}
\description{
This function allows you to visualize a palette
}
\examples{
\dontrun{
hex_vec <- c("#E69F00", "#56B4E9", "#009E73", "#F0E442", "#0072B2", "#D55E00",
"#CC79A7", "#666666", "#AD7700", "#1C91D4", "#007756", "#D5C711",
"#005685", "#A04700")
palette_viz(col_vec = hex_vec)
}

}
