% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_spatial.R
\name{qc_st_plots}
\alias{qc_st_plots}
\title{This function returns QC metric plots with Spatial Seurat objects.}
\usage{
qc_st_plots(
  se,
  nfeat = "nFeature_Spatial",
  ncount = "nCount_Spatial",
  slot = "counts",
  assay = "Spatial",
  percent.mito = NULL,
  percent.ribo = NULL
)
}
\arguments{
\item{se}{Spatial Seurat object we want to assess.}

\item{nfeat}{Character string indicating the name of the variable in the
metadata containing the number of genes.}

\item{ncount}{Character string indicating the name of the variable in the
metadata containing the number of UMIs.}

\item{slot}{slot from where we want to extract the count matrix.}

\item{assay}{assay from where we want to extract the count matrix.}

\item{percent.mito}{Character string indicating the name of the variable in
the metadata containing the mitochondrial percentage. Defaults to NULL in 
which case it computes it using the gollowing regex "^MT-".}

\item{percent.ribo}{Character string indicating the name of the variable in
the metadata containing the ribosomal percentage. Defaults to NULL in 
which case it computes it using the gollowing regex "^RPL|^RPS".}
}
\value{
List with the 4 QC plots
}
\description{
This function returns QC metric plots with Spatial Seurat objects.
}
\examples{
\dontrun{
library(Seurat)
library(SeuratData)
sp_obj <- SeuratData::LoadData(ds = "stxBrain", type = "posterior1")
qc_st_plots(
  se = sp_obj,
  nfeat = "nFeature_Spatial",
  ncount = "nCount_Spatial",
  slot = "counts",
  assay = "Spatial") \%>\%
cowplot::plot_grid(
  plotlist = .,
  align = "hv",
  axis = "trbl")
}

}
