% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_spatial.R
\name{ribo_pct_hist_st}
\alias{ribo_pct_hist_st}
\title{This function returns the ribosomal \% histogram of the Seurat object.}
\usage{
ribo_pct_hist_st(se, ribo_percent = "percent.ribo", sample_id = "orig.ident")
}
\arguments{
\item{se}{Spatial Seurat object we want to assess.}

\item{ribo_percent}{Character string indicating the name of the variable in
the metadata containing the ribosomal \% for each spot.}

\item{sample_id}{Feature in metadata by which to identify the sample if the
object is merged and contains several samples.}
}
\value{
Histogram with the Ribosomal \% distribution by sample.
}
\description{
This function returns the ribosomal \% histogram of the Seurat object.
}
\examples{
\dontrun{
library(Seurat)
library(SeuratData)
sp_obj <- SeuratData::LoadData(ds = "stxBrain", type = "posterior1")
ribo_pct_hist_st(se = sp_obj,
                 ribo_percent = "percent.ribo",
                 sample_id = "orig.ident")
}

}
